% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{json_validator}
\alias{json_validator}
\title{Create a json validator}
\usage{
json_validator(schema)
}
\arguments{
\item{schema}{Contents of the json schema, or a filename
containing a schema.}
}
\description{
Create a validator that can validate multiple json files.
}

